PracticeCode

+++ Day2Codehere/Day13codehere(Recursion)/CodeNumber3.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main() {
     int n;
     cin >> n; 
-    vector<int> nums(n);
+    vector<long long> nums(n);
 
     for (int i = 0; i < n; i++) {
         cin >> nums[i];
